+++ platformio/builder/scripts/baseavr.py
@@ -1,4 +1,4 @@
-# Copyright 2014-2015 Ivan Kravets <me@ikravets.com>
+# Copyright 2014-2016 Ivan Kravets <me@ikravets.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -42,10 +42,7 @@ env.Replace(
         "-mmcu=$BOARD_MCU"
     ],
 
-    CFLAGS=["-std=gnu11"],
-
     CXXFLAGS=[
-        "-std=gnu++11",
         "-fno-exceptions",
         "-fno-threadsafe-statics"
     ],