+++ platformio/builder/scripts/teensy.py
@@ -1,4 +1,4 @@
-# Copyright 2014-2015 Ivan Kravets <me@ikravets.com>
+# Copyright 2014-2016 Ivan Kravets <me@ikravets.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -28,7 +28,12 @@ if env.get("BOARD_OPTIONS", {}).get("build", {}).get("core") == "teensy":
 elif env.get("BOARD_OPTIONS", {}).get("build", {}).get("core") == "teensy3":
     SConscript(env.subst(join("$PIOBUILDER_DIR", "scripts", "basearm.py")))
     env.Append(
-        LINKFLAGS=["-Wl,--defsym=__rtc_localtime=$UNIX_TIME"]
+        LINKFLAGS=[
+            "-Wl,--defsym=__rtc_localtime=$UNIX_TIME"
+        ],
+        CPPFLAGS=[
+            "-fsingle-precision-constant"
+        ]
     )
 
 env.Append(