+++ platformio/commands/update.py
@@ -1,4 +1,4 @@
-# Copyright 2014-2015 Ivan Kravets <me@ikravets.com>
+# Copyright 2014-2016 Ivan Kravets <me@ikravets.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -24,4 +24,5 @@ from platformio.commands.platforms import \
 @click.pass_context
 def cli(ctx):
     ctx.invoke(cmd_platforms_update)
+    click.echo()
     ctx.invoke(cmd_lib_update)