+++ platformio/ide/projectgenerator.py
@@ -1,4 +1,4 @@
-# Copyright 2014-2015 Ivan Kravets <me@ikravets.com>
+# Copyright 2014-2016 Ivan Kravets <me@ikravets.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -74,9 +74,9 @@ class ProjectGenerator(object):
                 "\n".join([result['out'], result['err']]))
 
         output = result['out']
-        start_index = output.index('\n{"')
+        start_index = output.index('{"')
         stop_index = output.rindex('}')
-        data = json.loads(output[start_index + 1:stop_index + 1])
+        data = json.loads(output[start_index:stop_index + 1])
 
         return data
 
@@ -90,14 +90,6 @@ class ProjectGenerator(object):
                 for f in files:
                     result.append(relpath(join(root, f)))
         return result
-
-    @staticmethod
-    def get_main_src_file(src_files):
-        for f in src_files:
-            for ext in ("c", "cpp"):
-                if f.endswith(".%s" % ext):
-                    return f
-        return None
 
     def get_tpls(self):
         tpls = []
@@ -132,9 +124,9 @@ class ProjectGenerator(object):
 
     def _gather_tplvars(self):
         src_files = self.get_src_files()
-        main_src_file = self.get_main_src_file(src_files)
 
-        if not main_src_file and self.ide == "clion":
+        if (not any([f.endswith((".c", ".cpp")) for f in src_files]) and
+                self.ide == "clion"):
             click.secho(
                 "Warning! Can not find main source file (*.c, *.cpp). So, "
                 "code auto-completion is disabled. Please add source files "
@@ -147,7 +139,6 @@ class ProjectGenerator(object):
         self._tplvars.update({
             "project_name": self.get_project_name(),
             "src_files": src_files,
-            "main_src_file": main_src_file,
             "user_home_dir": abspath(expanduser("~")),
             "project_dir": self.project_dir,
             "systype": util.get_systype(),