+++ tests/commands/test_boards.py
@@ -1,4 +1,4 @@
-# Copyright 2014-2015 Ivan Kravets <me@ikravets.com>
+# Copyright 2014-2016 Ivan Kravets <me@ikravets.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,12 +13,9 @@
 # limitations under the License.
 
 import json
-from os.path import isfile, join
 
 from platformio import util
 from platformio.commands.boards import cli as cmd_boards
-from platformio.commands.platforms import \
-    platforms_install as cmd_platforms_install
 from platformio.commands.platforms import \
     platforms_search as cmd_platforms_search
 
@@ -52,21 +49,3 @@ def test_board_options(platformio_setup, clirunner, validate_cliresult):
     for _, opts in util.get_boards().iteritems():
         assert required_opts.issubset(set(opts))
         assert opts['platform'] in platforms
-
-
-def test_board_ldscripts(platformio_setup, clirunner, validate_cliresult):
-    result = clirunner.invoke(
-        cmd_platforms_install, [
-            "ststm32",
-            "--skip-default-package",
-            "--with-package=ldscripts"
-        ])
-    validate_cliresult(result)
-    ldscripts_path = join(util.get_home_dir(), "packages", "ldscripts")
-    for type_, opts in util.get_boards().iteritems():
-        if opts['build'].get("ldscript"):
-            frameworks = opts['frameworks']
-            if (any(fw in frameworks for fw in ["libopencm3", "mbed"]) or
-                    type_ in ("rfduino", )):
-                continue
-            assert isfile(join(ldscripts_path, opts['build'].get("ldscript")))